+++ ACBrFramework.Net/ACBrFramework.Net/ACBrPAF/Registros/ACBrPAFRegistroR7.cs
@@ -8,11 +8,11 @@ namespace ACBrFramework
 
 		public string MP { get; set; }
 
-		public double VL_PAGTO { get; set; }
+		public decimal VL_PAGTO { get; set; }
 
 		public string IND_EST { get; set; }
 
-		public double VL_EST { get; set; }
+		public decimal VL_EST { get; set; }
 
 		public bool RegistroValido { get; set; }
 	}
